+++ cut_template_long.py
@@ -56,7 +56,7 @@ def sac_ch_time(st):
         tr.stats.sac.nzhour = t0.hour
         tr.stats.sac.nzmin = t0.minute
         tr.stats.sac.nzsec = t0.second
-        tr.stats.sac.nzmsec = t0.microsecond / 1e3
+        tr.stats.sac.nzmsec = int(t0.microsecond / 1e3)
     return st
 
 def cut_event_window(stream_paths, tp, ts, out_paths):
